Log long Mediaflux requests (#2120)
Since we are going to fetch data from Mediaflux more frequently now that
we fetch the list of projects straight from Mediaflux (issue #2037) and
soon also the project details (#2039) it will be good to have a sense of
how long these queries are taking in Mediaflux.

For now I am logging as a warning any query that takes longer than 3
seconds. We can tweak that value once we see what's a good benchmark to
use. We could also send to Honeybadger extremely slow requests, again,
once we know what "extreme" means in the Mediaflux sense.