Project Show page now displays metadata straight from Mediaflux (#2128)
Updated project show page (Show and Details) to fetch the project
metadata straight from Mediaflux.

Notice that I am performing most of the changes at the Presenter level
(rather than at the Model) because we need a valid user session to fetch
from Mediaflux and the presenter seems like a better place for this.

NOTE: There are other places in the code that still reference the
original project metadata stored in our Rails record (e.g. mailers) and
we should refactor that code too to fetch the data from Mediaflux but I
am trying to keep this PR as small as possible. See issue:
https://github.com/pulibrary/tigerdata-app/issues/2132 for details.

Closes #2039