Fixes bug in check for user_has_access for data users (#2164)
There was a bug in the presenter code that checks whether a user has
access to a project, the code was treating the `data_users` as a list of
`uid`s but that list is now a list of `User`s. I probably introduced
this bug when I refactored the code that fetches data straight from
Mediaflux.

This PR fixes the issue and adds a test for it.