Include a custom app when logging in to Mediaflux (#2214)
Include a custom `app` parameter when logging in to Mediaflux. This
allows us to better track sessions that are coming in from the frontend
vs other sessions with the following command from aTerm:

```
system.session.all.describe :app "TDWEB"
```

which hopefully will help when troubleshooting the issue with the
licenses (#2189), although it is not clear to me how sessions and
licenses relate to each other.