Automatically log Mediaflux errors (#2215)
Automatically log to the Rails error log when calls to Mediaflux return
errors.

Before this PR errors were only logged if we called
`request.response_error` which we might or not do, but with this PR
errors will be logged as soon as the Mediaflux request is `resolved` and
that guarantees that errors will be logged at all times. I am not
sending the errors to Honeybadger because I have no idea how many errors
we will see in the logs. Once we see how our logs look like we could
decided what to send to Honeybadger.

Closes #2167