Improve logging when deleting asset.query.iterators (#2228)
When deleting an iterator we now ignore errors if the iterator is
missing. This is because Mediaflux automatically deletes iterators when
we have read all its contents which happens if the project has only a
handful of files and they are read all at once in the first iteration.
In that case we get the following error:

```
E, [2025-11-19T15:01:29.960006 #2438857] ERROR -- : [b6390527-4de0-4e5a-b5e3-f50585d45b5f]
Mediaflux error: arc.mf.server.Services$ExServiceError, call to service 'asset.query.iterator.destroy' failed:
The session does not have an iterator (id): 1
```

This is because Mediaflux automatically deletes the iterator as soon as
we read all the files from the query.