Better error handling in Project import  (#2294)
Don't stop processing import if there is an error in a project, instead
log the error and continue with other projects.

This is so that we can re-run the import in QA and have projects that
match what's on Mediaflux. We'll run this also in Staging (and perhaps
Production)


Related to #2209